test/bus: Break up dispatch test into three separate tests
This is really three separate test-cases: one for traditional
activation as a direct child process of the dbus-daemon, and two for
traditional activation (successful and failing) via the setuid
dbus-daemon-launch-helper on Unix.

The ones where activation succeeds are extremely slow, as a result of the
instrumentation for simulating malloc() failures combined with a large
number of memory operations, particularly when using AddressSanitizer.

Splitting up "OOM" tests like these has a disproportionately good impact
on the time they take, because the simulated malloc() failure
instrumentation repeats the entire test making the first malloc() fail,
then making the second malloc() fail, and so on. For allocation failures
in the second half of the test, this means we repeat the first half of
the test with no malloc() failures a very large number of times, which
is not a good use of time, because we already tested it successfully.

Even when not using the "OOM" instrumentation, splitting up these tests
lets them run in parallel, which is also a major time saving.

Needless to say, this speeds up testing considerably. On my modern
but unexceptional x86 laptop, in a typical debug build with Meson on
the 1.15.x branch, the old dispatch test took just over 21 minutes,
which drops to about 40 seconds each for the new normal-activation and
helper-activation tests (and for most of that time, they're running
in parallel, so the wall-clock time taken for the whole test suite is
somewhere around a minute).

In a debug build with Meson, gcc and AddressSanitizer on the 1.15.x
branch, the old dispatch test takes longer than my patience will allow,
and the new separate tests take about 5-6 minutes each. Reduce their
timeout accordingly, but not as far as the default for slow tests (5
minutes) to allow some headroom for AddressSanitizer or slower systems.

A side benefit of this is that it makes our testing more realistic.  Each
call to bus_dispatch_test_conf() creates a new BusContext, which has the
side-effect of saving the initial file descriptor rlimit and expanding the
soft limit to match the hard limit. When we call bus_dispatch_test_conf()
more than once per test process, the second and subsequent calls will save
the increased rlimit as though it had been the original, and restore it
before exec'ing subprocesses, resulting in a potentially extremely high
soft limit which means our debug instrumentation takes a very long time
to iterate through all possible fds (see dbus#527).

Signed-off-by: Simon McVittie <smcv@collabora.com>
(cherry picked from commit bef88fd5627f2d990915c9009982107a8e329ef5)
Mitigates: https://gitlab.freedesktop.org/dbus/dbus/-/issues/527