meson: Don't interpolate the result of configure_file() into a format()
configure_file() returns a file object, which was not intended to be a
valid parameter for format() (even though in practice it works the way
we wanted it to), causing newer Meson versions to report deprecation
warnings.

If the version of Meson is new enough, we can use the full_path()
method. Otherwise, we'll have to re-compute the output filename.

Signed-off-by: Simon McVittie <smcv@collabora.com>
(cherry picked from commit 872ce4d29a4c34b1ae8bb5d4e02fdc7a600a3075)