INSTALL, README: Simplify Meson vs. CMake recommendation
Now that 1.16.0 has been released and 1.15.x is EOL, it's misleading
to say that Meson is the recommended build system for Unix only on the
master branch: it's the recommended build system for Unix on the
dbus-1.16 branch, too.

This also avoids explicitly naming the master branch, which is likely
to get renamed to main.

Signed-off-by: Simon McVittie <smcv@collabora.com>