README.win, README.cygwin: Describe dbus4win, etc. as historical
The merge of Windows-specific dbus ports into the same git repository
as the original Unix-only implementation was a long time ago now,
and exists on several stable-branches, not just master.

This also avoids explicitly mentioning the name of the master branch,
which is likely to be renamed to main in future.

Signed-off-by: Simon McVittie <smcv@collabora.com>
(cherry picked from commit 33d48768062ddf446d1b108509ee10eec2871f32)